Rename files; split tests to different files
